pyproject.toml
[project]
name = "yt-bot"
version = "0.1.0"
description = "Simple bot to add fun to friends' server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "pynacl>=1.6.1",
    "python-dotenv>=1.2.1",
]
